requirements.txt
pytest
flask
dotenv
flask_sqlalchemy
psycopg2-binary